+++ changelog.md
@@ -1,5 +1,5 @@
 ## Changes for 20230708.0.0
 
-* Updated translations.
+* Added compatibility with nvda-2023.2.0 for settings dialog.